[DEV-13147] init progress on refactor
