+++ usaspending_api/etl/management/commands/load_query_to_delta.py
@@ -19,10 +19,10 @@ from usaspending_api.disaster.delta_models import (
     covid_faba_spending_load_sql_strings,
 )
 from usaspending_api.disaster.models import CovidFABASpending
-from usaspending_api.download.delta_models.account_download import (
-    ACCOUNT_DOWNLOAD_POSTGRES_COLUMNS,
-    account_download_create_sql_string,
-    account_download_load_sql_string,
+from usaspending_api.download.delta_models.award_financial_download import (
+    AWARD_FINANCIAL_DOWNLOAD_POSTGRES_COLUMNS,
+    award_financial_download_create_sql_string,
+    award_financial_download_load_sql_string,
 )
 from usaspending_api.download.delta_models.object_class_program_activity_download import (
     OBJECT_CLASS_PROGRAM_ACTIVITY_DOWNLOAD_POSTGRES_COLUMNS,
@@ -321,10 +321,10 @@ TABLE_SPEC = {
         "tsvectors": None,
         "postgres_partition_spec": None,
     },
-    "account_download": {
+    "award_financial_download": {
         "model": None,
         "is_from_broker": False,
-        "source_query": [account_download_load_sql_string],
+        "source_query": [award_financial_download_load_sql_string],
         "source_query_incremental": None,
         "source_database": None,
         "source_table": None,
@@ -334,10 +334,10 @@ TABLE_SPEC = {
         "partition_column": "financial_accounts_by_awards_id",
         "partition_column_type": "numeric",
         "is_partition_column_unique": False,
-        "delta_table_create_sql": account_download_create_sql_string,
-        "source_schema": ACCOUNT_DOWNLOAD_POSTGRES_COLUMNS,
+        "delta_table_create_sql": award_financial_download_create_sql_string,
+        "source_schema": AWARD_FINANCIAL_DOWNLOAD_POSTGRES_COLUMNS,
         "custom_schema": None,
-        "column_names": list(ACCOUNT_DOWNLOAD_POSTGRES_COLUMNS),
+        "column_names": list(AWARD_FINANCIAL_DOWNLOAD_POSTGRES_COLUMNS),
         "postgres_seq_name": None,
         "tsvectors": None,
         "postgres_partition_spec": None,
