+++ usaspending_api/etl/submission_loader_helpers/program_activities.py
@@ -26,7 +26,7 @@ def update_program_activities(submission_id):
                 create_date,
                 update_date
             )
-        select
+        SELECT
             program_activity_code,
             program_activity_name,
             reporting_fiscal_year,
